Adding dir/file count support and option to include hidden files
I added two fields, ndirs and nfiles, to the TreeItem class, as
well as two corresponding functions used to count the number of
files and directories that are children of a given TreeItem. This
makes it easy to give an output of the number of files/dirs at the
bottom of the pypree output.

I also added a -a/--all option to the CLI so that users can specify
if they want to include hidden files (i.e. those beggining with ".")
, which are now excluded from the output by default.