Implemented proper indentation/bar count based on parent/child positions
Now, the number of vertical bars on a given line corresponds exactly
with the tree command. This number is, for a given child node that
is at level two or deeper in the tree, a function of the number of
of times that a child's parent is the last child of that child's
grandparent.