Diff identical with tree. Closes #1.
It took me a bit to figure out specifically where I needed to add
non-breaking spaces versus "regular" spaces, but I did it.

I think the only thing diff should pick up now is if filenames have spaces
in them. It seems that the tree command adds a "\" to escape the
space in this case. I'm going to assume that any users of the program
choose sane filenames that don't include spaces (unlike me). :)